Create another one, under the first code one
First, the "Branches" button was pushed. This gave a view of all the branches under Master. (There was only one.) A "View #2" button looked inviting, and this gave a screen where "Add new file" was available. (No other screen had this.) The assumption is that this will allow a new file to be put as a branch out of that first branch, which was some code.